mqtt/pyproject.toml
[project]
name = "mqtt-smcp-server"
version = "0.1.0"
description = "SMCP-enabled MCP server for MQTT with secure credential injection"
authors = [
    { name = "SMCP Project" },
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["mqtt", "mcp", "smcp", "model-context-protocol"]
dependencies = [
    "mcp>=1.6.0",
    "paho-mqtt>=2.0.0",
    "smcp",
]

[project.scripts]
mqtt-smcp-server = "mqtt_smcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mqtt_smcp_server"]
